+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-
+
+### Changed
+
+-
+
 ## [0.5.0] - 2026-02-22
 
 ### Added
@@ -70,6 +78,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Operational local cache.
 - ScoreCalculator unit tests (>80% coverage)
 
-[unreleased]: https://github.com/meragix/cura/compare/cura-0.2.0...HEAD
+[unreleased]: https://github.com/meragix/cura/compare/cura-0.5.0...HEAD
+[0.5.0]: https://github.com/meragix/cura/releases/tag/cura-0.5.0
+[0.4.0]: https://github.com/meragix/cura/releases/tag/cura-0.4.0
+[0.3.0]: https://github.com/meragix/cura/releases/tag/cura-0.3.0
 [0.2.0]: https://github.com/meragix/cura/releases/tag/cura-0.2.0
 [0.1.0]: https://github.com/meragix/cura/releases/tag/cura-0.1.0